requirements.txt
PySide6>=6.6.0,<7.0.0
GitPython>=3.1.30,<4.0.0
google-generativeai>=0.4.0,<1.0.0
python-dotenv>=1.0.0,<2.0.0
PyYAML>=6.0,<7.0.0        